requirements.txt
scikit-learn
pandas
numpy
nltk
joblib
datasets
flask
flask-cors
wordcloud
matplotlib
python-whois
dnspython
requests
python-dotenv